+++ frontend/public/index.html
@@ -5,7 +5,18 @@
 		<link rel="icon" href="%PUBLIC_URL%/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<meta name="theme-color" content="#000000" />
-		<meta name="description" content="Machine Learning made equal" />
+		<meta
+			name="description"
+			content="MLforAll's mission is to make machine learning education access for everyone. It takes away the code and lets you worry about asking the big questions."
+		/>
+		<meta property="og:type" content="website" />
+		<meta
+			name="og:title"
+			property="og:title"
+			content="MLforALL, Machine Learning made easier for everyone"
+		/>
+
+		<title>Machine Learning for All</title>
 		<link rel="apple-touch-icon" href="%PUBLIC_URL%/favicon.png" />
 		<!--
       manifest.json provides metadata used when your web app is installed on a
@@ -36,7 +47,6 @@
 		/>
 		<!-- Compiled and minified JavaScript -->
 		<script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
-		<title>ML For All</title>
 	</head>
 	<body>
 		<!-- The core Firebase JS SDK is always required and must be listed first -->
