+++ frontend/public/manifest.json
@@ -1,6 +1,6 @@
 {
 	"short_name": "MLforALL",
-	"name": "MLforAll: Classification Models made Equal",
+	"name": "Machine Learning for All",
 	"icons": [
 		{
 			"src": "favicon.ico",
