+++ frontend/src/styling/landing.css
