+++ frontend/src/styling/projectdetails.css
